# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "OpenDispatch",
    "description": "The REST half of the OpenDispatch API contract. SignalR board events and the offline-sync payloads are not describable here; they live in @opendispatch/contracts.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "operationId": "CreateJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateJobResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "ListJobs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPageResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "GetJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/jobs/{id}/status": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "operationId": "ChangeJobStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeJobStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/jobs/{id}/assign": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "operationId": "AssignJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignJobResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/sync/push": {
      "post": {
        "tags": [
          "Sync"
        ],
        "operationId": "PushSyncOps",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/sync/pull": {
      "get": {
        "tags": [
          "Sync"
        ],
        "operationId": "PullSyncChanges",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AssignJobRequest": {
        "required": [
          "technicianId",
          "scheduledStart"
        ],
        "type": "object",
        "properties": {
          "technicianId": {
            "type": "string",
            "description": "Whose day it goes on.",
            "format": "uuid"
          },
          "scheduledStart": {
            "type": "string",
            "description": "When they are planned to arrive.",
            "format": "date-time"
          }
        },
        "description": "The body of `POST /jobs/{id}/assign`."
      },
      "AssignJobResponse": {
        "required": [
          "assignmentId"
        ],
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string",
            "description": "The identity of the stop it created or moved — the board has just drawn it.",
            "format": "uuid"
          }
        },
        "description": "What a successful `POST /jobs/{id}/assign` returns."
      },
...
      "ChangeJobStatusRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "description": "Where the job should be.",
            "$ref": "#/components/schemas/JobStatus"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "When the work actually finished. Only read when Status is\nJobStatus.Completed; when it is omitted the server asks its own clock.",
            "format": "date-time"
          }
        },
        "description": "The body of `POST /jobs/{id}/status`."
      },
...
      "CreateJobRequest": {
        "required": [
          "customerId",
          "locationId",
          "requiredSkill",
          "priority",
          "windowStart",
          "windowEnd",
          "estimatedDuration"
        ],
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "description": "Whose work it is.",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "description": "Which of that customer's service locations it happens at.",
            "format": "uuid"
          },
          "requiredSkill": {
            "type": "string",
            "description": "The skill a technician must have to take it."
          },
          "priority": {
            "description": "How badly it needs doing.",
            "$ref": "#/components/schemas/JobPriority"
          },
          "windowStart": {
            "type": "string",
            "description": "When the promised window opens.",
            "format": "date-time"
          },
          "windowEnd": {
            "type": "string",
            "description": "When the promised window closes.",
            "format": "date-time"
          },
          "estimatedDuration": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string",
            "description": "How long the work should take once a technician is on site."
          }
        },
        "description": "The body of `POST /jobs`."
      },
      "CreateJobResponse": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The identity of the job it created.",
            "format": "uuid"
          }
        },
        "description": "What a successful `POST /jobs` returns."
      },
...
      "JobPageResponse": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "description": "The jobs on this page, soonest-promised first."
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "How many the tenant has altogether.",
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Which page this is, counting from one.",
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "How many this page was allowed to hold.",
            "format": "int32"
          }
        },
        "description": "One page of jobs. The body of `GET /jobs?page={n}&amp;pageSize={n}`."
      },
      "JobPriority": {
        "enum": [
          "Low",
          "Normal",
          "High",
          "Emergency"
        ],
        "description": "How badly a job needs doing, as the clients see it."
      },
      "JobResponse": {
        "required": [
          "id",
          "customerId",
          "locationId",
          "latitude",
          "longitude",
          "requiredSkill",
          "priority",
          "windowStart",
          "windowEnd",
          "estimatedDuration",
          "status",
          "notes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Its identity.",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "description": "Whose work it is.",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "description": "Which of that customer's service locations it happens at.",
            "format": "uuid"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Where it is, in decimal degrees.",
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Where it is, in decimal degrees.",
            "format": "double"
          },
          "requiredSkill": {
            "type": "string",
            "description": "The skill a technician must have to take it."
          },
          "priority": {
            "description": "How badly it needs doing.",
            "$ref": "#/components/schemas/JobPriority"
          },
          "windowStart": {
            "type": "string",
            "description": "When the promised window opens.",
            "format": "date-time"
          },
          "windowEnd": {
            "type": "string",
            "description": "When the promised window closes.",
            "format": "date-time"
          },
          "estimatedDuration": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string",
            "description": "How long the work should take once a technician is on site."
          },
          "status": {
            "description": "How far through its life the job is.",
            "$ref": "#/components/schemas/JobStatus"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ],
            "description": "What a technician wrote about it, or `null` if nobody has."
          }
        },
        "description": "One job, as the clients see it: the demand, and nothing about the plan."
      },
      "JobStatus": {
        "enum": [
          "Unscheduled",
          "Scheduled",
          "Dispatched",
          "EnRoute",
          "InProgress",
          "Completed",
          "Invoiced",
          "Paid",
          "Cancelled"
        ],
        "description": "Where a job has got to in its life, as the clients see it."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SyncConflict": {
        "required": [
          "opId",
          "reason",
          "message"
        ],
        "type": "object",
        "properties": {
          "opId": {
            "type": "string",
            "description": "Which operation, by the id the device gave it.",
            "format": "uuid"
          },
          "reason": {
            "description": "What kind of refusal it was, for the client to branch on.",
            "$ref": "#/components/schemas/SyncConflictReason"
          },
          "message": {
            "type": "string",
            "description": "Why, in terms fit to show the technician holding the phone."
          }
        },
        "description": "An operation the server would not apply, and why."
      },
...
      "SyncPullResponse": {
        "required": [
          "changes",
          "cursor",
          "hasMore"
        ],
        "type": "object",
        "properties": {
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncChange"
            },
            "description": "The entities that moved, as the server now holds them."
          },
          "cursor": {
            "type": "string",
            "description": "The watermark to ask with next time. Opaque — the shape of the server's bookmark is not\npart of the contract."
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether the server stopped early and has more waiting. When `true`, pull again\nwith Cursor as soon as this page is applied; when `false`,\nthe device is up to date as of Cursor."
          }
        },
        "description": "Everything that changed in a device's world since the cursor it asked with, and where it\nnow stands. The body of `GET /sync/pull?since={cursor}`."
      },
      "SyncPushRequest": {
        "required": [
          "ops"
        ],
        "type": "object",
        "properties": {
          "ops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncOp"
            },
            "description": "The operations, oldest first."
          }
        },
        "description": "A device emptying its queue: everything it did since it last got through, in the order it\ndid it. The body of `POST /sync/push`."
      },
      "SyncPushResponse": {
        "required": [
          "applied",
          "conflicts",
          "cursor"
        ],
        "type": "object",
        "properties": {
          "applied": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The ops that landed, by id. An op the server had already applied is reported here too:\nfrom the device's point of view a re-sent operation that is already in effect succeeded,\nand telling it otherwise would leave it queueing the op forever."
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncConflict"
            },
            "description": "The ops that were refused, each with a reason fit to show a technician."
          },
          "cursor": {
            "type": "string",
            "description": "Where the device now stands in the change stream. Opaque — it is the server's watermark,\nand a client that parses it is reading something it was not promised."
          }
        },
        "description": "What the server made of a pushed batch. Authoritative: the device rebases onto this rather\nthan keeping its own opinion (Document 2 §10)."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "The token POST /auth/login returns, sent as \"Authorization: Bearer <token>\".",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}